+++ packages/ag-grid-theme/stories/examples/RowGrouping.tsx
@@ -1,28 +1,24 @@
-import { StackLayout } from "@salt-ds/core";
 import { AgGridReact, AgGridReactProps } from "ag-grid-react";
+import { useAgGridThemeSwitcher } from "../dependencies/ThemeSwitcher";
 import dataGridExampleData from "../dependencies/dataGridExampleData";
 import dataGridExampleRowGrouping from "../dependencies/dataGridExampleRowGrouping";
 import { useAgGridHelpers } from "../dependencies/useAgGridHelpers";
-import { useAgGridThemeSwitcher } from "../dependencies/ThemeSwitcher";
 
 const RowGrouping = (props: AgGridReactProps) => {
-  const { switcher, themeName } = useAgGridThemeSwitcher();
+  const { themeName } = useAgGridThemeSwitcher();
   const { agGridProps, containerProps } = useAgGridHelpers({
     agThemeName: `ag-theme-${themeName}`,
   });
 
   return (
-    <StackLayout gap={4}>
-      {switcher}
-      <div {...containerProps}>
-        <AgGridReact
-          columnDefs={dataGridExampleRowGrouping}
-          rowData={dataGridExampleData}
-          {...agGridProps}
-          {...props}
-        />
-      </div>
-    </StackLayout>
+    <div {...containerProps}>
+      <AgGridReact
+        columnDefs={dataGridExampleRowGrouping}
+        rowData={dataGridExampleData}
+        {...agGridProps}
+        {...props}
+      />
+    </div>
   );
 };