+++ packages/ag-grid-theme/stories/examples/WrappedHeader.tsx
@@ -1,15 +1,9 @@
-import { ChangeEvent, useEffect, useState } from "react";
+import { Checkbox, FlexItem, StackLayout, useDensity } from "@salt-ds/core";
 import { AgGridReact, AgGridReactProps } from "ag-grid-react";
-import {
-  StackLayout,
-  FlexItem,
-  FlexLayout,
-  Checkbox,
-  useDensity,
-} from "@salt-ds/core";
+import { ChangeEvent, useEffect, useState } from "react";
+import { useAgGridThemeSwitcher } from "../dependencies/ThemeSwitcher";
 import dataGridExampleColumnsWrap from "../dependencies/dataGridExampleColumnsWrap";
 import dataGridExampleData from "../dependencies/dataGridExampleData";
-import { useAgGridThemeSwitcher } from "../dependencies/ThemeSwitcher";
 import { useAgGridHelpers } from "../dependencies/useAgGridHelpers";
 
 const statusBar = {
@@ -26,7 +20,7 @@ const statusBar = {
 
 const WrappedHeader = (props: AgGridReactProps) => {
   const [compact, setCompact] = useState(false);
-  const { switcher, themeName } = useAgGridThemeSwitcher();
+  const { themeName } = useAgGridThemeSwitcher();
   const { api, agGridProps, containerProps, isGridReady } = useAgGridHelpers({
     agThemeName: `ag-theme-${themeName}`,
     compact,
@@ -47,17 +41,14 @@ const WrappedHeader = (props: AgGridReactProps) => {
 
   return (
     <StackLayout gap={4}>
-      <FlexLayout direction="row">
-        {switcher}
-        <FlexItem>
-          <Checkbox
-            checked={compact && density === "high"}
-            label="Compact (for high density only)"
-            onChange={handleCompactChange}
-            disabled={density !== "high"}
-          />
-        </FlexItem>
-      </FlexLayout>
+      <FlexItem>
+        <Checkbox
+          checked={compact && density === "high"}
+          label="Compact (for high density only)"
+          onChange={handleCompactChange}
+          disabled={density !== "high"}
+        />
+      </FlexItem>
       <div {...containerProps} style={{ height: "400px", width: "500px" }}>
         <AgGridReact
           columnDefs={dataGridExampleColumnsWrap}