+++ packages/core/src/breakpoints/index.ts
@@ -1,1 +1,2 @@
 export * from "./Breakpoints";
+export * from "./BreakpointProvider";