+++ packages/core/src/menu/MenuTriggerContext.ts
@@ -1,0 +1,16 @@
+import { useContext } from "react";
+import { createContext } from "../utils";
+
+export interface MenuTriggerContextValue {
+  triggersSubmenu: boolean;
+  blurActive: boolean;
+}
+
+export const MenuTriggerContext = createContext<MenuTriggerContextValue>(
+  "MenuTriggerContext",
+  { triggersSubmenu: false, blurActive: false }
+);
+
+export function useIsMenuTrigger() {
+  return useContext(MenuTriggerContext);
+}