+++ packages/icons/src/components/CopySolid.tsx
@@ -15,8 +15,12 @@ export const CopySolidIcon = forwardRef<SVGSVGElement, CopySolidIconProps>(
         ref={ref}
         {...props}
       >
-        <path d="M8.707 0H4v9h7V2.293L8.707 0zM7 1h1v2h2v1H7V1z" />
-        <path d="M8 10H3V3H1v9h7v-2z" />
+        <path
+          fillRule="evenodd"
+          d="M8.707 0H4v9h7V2.293L8.707 0ZM7 1h1v2h2v1H7V1Z"
+          clipRule="evenodd"
+        />
+        <path d="M8 10H3V3H1v9h7v-2Z" />
       </Icon>
     );
   }