+++ packages/icons/src/components/PlaceIn.tsx
@@ -15,8 +15,8 @@ export const PlaceInIcon = forwardRef<SVGSVGElement, PlaceInIconProps>(
         ref={ref}
         {...props}
       >
-        <path d="M0 12h12V6h-1v5H1V1h5V0H0v12z" />
-        <path d="M10.939 0 12 1.061 9.025 4.036 10.489 5.5h-4v-4l1.475 1.475L10.939 0z" />
+        <path d="M0 12h12V6h-1v5H1V1h5V0H0v12Z" />
+        <path d="M10.94 0 12 1.06 8.525 4.537 9.99 6h-4V2l1.474 1.475L10.94 0Z" />
       </Icon>
     );
   }