+++ packages/icons/src/components/PrintSolid.tsx
@@ -15,7 +15,11 @@ export const PrintSolidIcon = forwardRef<SVGSVGElement, PrintSolidIconProps>(
         ref={ref}
         {...props}
       >
-        <path d="M2 0h8v2h2v6h-2v4H2V8H0V2h2V0zm7 2V1H3v1h6zm1 3H2v2h1V6h6v1h1V5zm-2 5V9H4v1h4zm0-3H4v1h4V7z" />
+        <path
+          fillRule="evenodd"
+          d="M2 0v2H0v6h2v4h8V8h2V2h-2V0H2Zm7 2V1H3v1h6Zm1 3H2v2h1V6h6v1h1V5Zm-2 5V9H4v1h4Zm0-3v1H4V7h4Z"
+          clipRule="evenodd"
+        />
       </Icon>
     );
   }