+++ packages/icons/src/components/VolumeDown.tsx
@@ -15,7 +15,7 @@ export const VolumeDownIcon = forwardRef<SVGSVGElement, VolumeDownIconProps>(
         ref={ref}
         {...props}
       >
-        <path d="M3 4H0v4h3l3 3V1L3 4zm4 4.829a3.001 3.001 0 0 0 0-5.658V8.83z" />
+        <path d="M3 4H0v4h3l3 3V1L3 4Zm4 4.83a3.001 3.001 0 0 0 0-5.66v5.66Z" />
       </Icon>
     );
   }